requirements.txt
requests
datetime
python-dateutil
json_logging
slackclient
redis

pylint

boto3
